Add support for dynamic downsampling.
Queries can now specify a downsampling function as well as an interval.
This is useful for, say, plotting the 5m average over a long period of
time, where without downsampling the graph would be overcrowded due to
the sheer number of data points.  It's also useful to smooth curves that
are very jittery (e.g. 'max' of %CPU used over 10m).

This is done by adding a new iterator for Span that takes care of
downsampling the data before it's returned to the SpanGroup.

Change-Id: I01c63869e1e5982d0638cb184a8e0307b164d126