Make it possible to delete rows as they are scanned.
The `scan' tool can also delete data as it's being scanned, thanks to a new
`--delete' flag.  This is useful for removing bogus data from the TSDB.

The `scan' tool now also has a `--import' flag to dump the data in a format
suitable for the `import' tool, instead of using the default format which is
closer to HBase's representation of the data.

Change-Id: I71c22be3e0432e3f169c559b96d4c2b685968968