Add a query UI through the HTTP interface.
The UI is made with GWT and uses JSON to get various tidbits of data on the
client side, instead of relying on page generation on the server side.

How to test the UI:
  1. Run:
       ant gwtdev
     This starts GWT in "hosted" mode (i.e. the code in your browser is in
     fact Java executing in a JVM, instead of JavaScript, it's magic!)
  2. Background the `ant' process, or use another terminal:
       ant gwttsd
     This starts the TSD to serve the pages to your browser and use GWT
     in hosted mode.
  3. In the `GWT Development Mode' window started by the previous `ant'
     command, click "Launch Default Browser".
  4. Profit!

Deployment: still need to sort out the details.  Right now one can run
  ant gwtc
  ant gwttsd
And it'll work, but the TSDB's .jar is far from self-contained, so need to
sort this out package inside the various static resources including those
generated by the GWT compiler.

Change-Id: I936a29db519c1551c868724199726e1c5b12edf4