Switch from HBase's traditional client to my async client.
This is the first step of refactoring: the code uses the async client in
a sync / blocking way, which is stupid but requires fewer code changes.
The next step of the refactoring will be to change the code further in
order to take advantage of the asynchronous API to never block Netty IO
threads when interacting with HBase.

Change-Id: Ibea417d00f5e9bf25feea3368a3260f1b2f98dc1