Properly ignore data points that are read out of order from HBase.
This happens when overlapping rows are erroneously created.  The code
attempted to ignore them but it was leaving its data structure in an
inconsistent state.  We now keep a reference to the old state so we
can revert and abort when encountering a row that overlaps with one
we read previously.

Change-Id: Ib667c29e6584b901ee86ffa528d76a8caf2f921d