Implement /logs?level=LVL[&logger=PATH] in the HTTP API.
This allows users to change the logging level (verbosity) of the TSD at
runtime.  LVL can be debug, info, etc.  PATH should be the name of a
package (e.g, net.opentsdb.tsd) in which case the change will affect
all classes in this package, or the fully qualified name of a class
(e.g, net.opentsdb.tsd.LogsRpc).  If no `logger' argument is specified,
the change is applied on all loggers.

Change-Id: I30f188321213486c34cb4343a7d2d1059fb279ba