Put the stack traces in the logs RPC responses.
This change also adds proper escaping of log messages, which is
particularly important for JSON responses.

A new CSS style ("fwf" = fixed width font) is added.  Unfortunately GWT
overrides the font in each <td> in the FlexTable, but at least the text
is now wrapped around properly.

In addition, we also highlight the warning and error messages
differently in the web UI, so they stand out better.

Change-Id: I9ce4ee61f36058bd15164dd040e1b0a2b8d52ba7