Make all HTTP endpoints send return JSON-friendly error responses.
When generating an error response (say, 400, 404, 500), honor the
`json' query string parameter, which indicates that the client wants
a JSON response.  This was only done by /q, but is now automagically
done for all endpoints.

In the UI, take advantage of the JSON encoded error responses to
show more meaningful error messages than just "Internal Server Error"
and such.  This allows the UI to show stack traces from the server side
with detailed error messages.

Change-Id: I2af1a02341850e176daf4fff966999fc652701b6