Make all HTTP endpoints send return PNG-friendly error responses.
Similarly to the previous commit, error responses (say, 400, 404, 500),
now honor the `png' query string parameter, which indicates that the
client wants a PNG image as a response.  This was only handled by /q,
but is now automagically done for all endpoints.

Among other things, this helps report proper error messages to pages
that embed an image from the TSD without first doing an AJAX call to
see if there's an error.

Change-Id: Idb44d3ef78152cfb98c929e355c8f615f893b30f