Fix another off-by-one in a comparison.
A RowSeq can hold up to 2^12 = 4096 data points (one every second for
4096 consecutive seconds).  So the maximum time delta we can allow is
4095 not 4096, the comparison should have been strict.

This bug was causing the TSD to sometimes discard data points when
reading from HBase (the write path was unaffected).  When this was
happening, the TSD was logging an ERROR message like this:
  RowSeq: new timestamp = 1290564600 is < previous=1290568395 in addRow
  with kv=KeyValue(...)
The error is misleading as the new timestamp wasn't the one given in the
error message, it was that value + 4096 (so in the example above, the
data point the TSD is rejecting is at timestamp = 1290568696) because of
the 12-bit integer overflow caused by the off-by-one.

Change-Id: Iec35e08c66a6f7050e248f6610e646f7f411f0ae