Query code: properly look before/after the start/end dates.
When handling a query, we always need to start scanning a little bit
before the start date and stop scanning a little bit after the end date.
This is required to do proper rate computation or downsampling around
the "edges" of the graph.  Unfortunately, even though the scanner was
properly getting the necessary data points from HBase, when they were
passed to the SpanGroup, any data point that wasn't strictly inside the
start/end dates of the query would get discarded.  We now instead tell
the SpanGroup to use the same adjusted start/end dates as we use for
the scanner.  Gnuplot will not render any data point outside of the time
range we request anyway.

This prevents artificial effects where the graph appears to be tailing
off near the edges.

Change-Id: If952c42a734c015facaf6d9190458d5555f9d1d2