Bug fix: make queries twice faster by properly using the cache.
Right before the open-source release, the code was rebased to
retroactively add the `&png' parameter, however the code that
computes the cache key wasn't updated accordingly.  Because of
that, any query coming from the UI would actually be executed
twice (and would thus take twice the time to complete).  The
first query would come in with the `&json' suffix, and the PNG
would be rendered and cached.  Then the UI would request the
`&png' URL, but this URL was hashed to a different cache key,
so another PNG was generated (which led to another HBase scan).

Change-Id: I6e38f21d0ebc8252d931ba411bdabf1801a77824