Don't attempt to send an HTTP response on a closed channel.
If a user closes their browser window before the TSD starts sending the
response to the last query of this user, the browser will typically
close the connection instead of waiting for a reply it won't use (since
the window was closed).  This generates spurious exceptions that cause
spurious warnings in the logs ("Attempt to write to closed channel").

Change-Id: I8e65232eff5012b52a61e6fd897fc4a67a1d7f5b