Fix caching of requests with relative end time.
The logic to determine whether a cache file was stale was wrong when
the end time was relative.  More specifically, as long as the `end'
parameter was given, the cached file wouldn't be regenerated.  Also
the mtime of the cached file was compared against the end time of the
request instead of being compared to the current time to determine how
old the file is.

Change-Id: I77b810b26566f6b1fa9fc3bfffc1b862d16aa660