Fix a bug in the aggregation code.
Background:
  In the aggregation code, we maintain multiple iterators on the time
  series to aggregate.  Every time we read an aggregated value, the code
  looks at all the iterators and advances the one that has the smallest
  timestamp.  When multiple iterators are at the same timestamp, they're
  all advanced at the same time.

The bug: when multiple iterators were advanced at the same time, the
code was improperly handling the case where any of the iterators reached
the end of their series except for the first one.  Because of this,
stale data wasn't removed from the internal data structure of the
aggregator and the stale data was erroneously aggregated, leading to
weird graphs with increasingly high values.

Change-Id: Ieb41c377d5311c79fc28608caf614f51a4a4360f