Add unit tests for `UniqueId.suggest()'.
Unfortunately these tests wouldn't catch the subtle performance bug
fixed by 1882ff04a93ce08fa95c6ae49385cd8b51494ec6, because the bug
was that any suggestion that was scanned was simply re-entered in
the caches even if it was already there, and there's no easy way to
find about this invisible side effect.  But at least there are tests
for this part of the API now.

Change-Id: Ie152d4f54351945bd6b66ba952def41b5199685e