Fix a bug in the server-side filter code.
Any UID that contained with the byte 0x5C (backslash `\') was
erroneously escaped in the scanner's filter.  There is no escaping
possible inside a \Q...\E sequence.  The only tricky case to handle
is when there's a literal \E inside the sequence.

What this means is that it was impossible to make a query that involved
such a UID.  For instance if you queried some metric for host=web42 and
it so happened that the UID for `web42' was [0, 0, 0x5C] then the query
would return nothing.

This closes #25.

Change-Id: Ide82b2ad89c89168505102a7b1f9c7ea7f076c13