Increase the maximum time span of a row from 10 to 60 minutes.
This means that now the TSD will store 1h worth of data per row.
Be careful when rolling out this change.  All TSDs need to be in
agreement as far as this value is concerned to avoid creating
"overlapping" rows.  The recommended way to roll out this change
is to do a rolling restart of TSDs within the first 10 minutes
after the beginning of a new hour.

This change is backwards and forward compatible.

This change also fixes a couple bugs that were lurking in
IncomingDataPoints due to Java's stupid integer promotion
when doing bit-operations on `short' integers!@#$%^

Change-Id: I66602348080dc5d58d3f7b0bc144d1a713bde607