Simplify the write path.
The idea behind IncomingDataPoints never really took off.  This class
was originally supposed to allow the TSD to maintain an in-memory cache
of "recent" data that had been written to HBase.  This in-memory cache
was never implemented and cannot actually be easily implemented because
we can no longer guarantee that a TSD sees all the data points going to
a particular row.  The idea was that this in-memory cache could be used
to serve queries on recent data and to ride over HBase outages (i.e.
serve recent data even if HBase was down).  I don't see TSD doing this
any time soon, so this change starts to bypass IncomingDataPoints, which
does some complicated things that aren't actually necessary.

Change-Id: Ifef6227e6f1d2766b14fe5b5701ece62ea5973c4