Implement the --fix flag for the "fsck" command.
This helps automatically fix out-of-order data problems.
Depending on the exact nature of the problem at hand, the
behavior implemented by --fix might not always be The Right
Thing.  With --fix, data scanned last can always overwrite
data scanner earlier, but that's not necessarily what the
user wants, it depends.

Change-Id: I0e68fe32af98f28889e06836289d7a0fbb32c36a