Add a new runtime exception for malformed data.
This is better than using AssertionError because this class of errors
isn't caused by an invariant being violated, rather by data being not
what we expect (malformed or corrupted).  In the TSD, this will lead
to an Internal Server Error, instead of killing the thread without
giving feedback to the remote client.

Change-Id: I611b565623468497aa701bd878dec16d6ffa44ca