Improved schema: compact the data in HBase.
This is a major change.  New data points written by this code aren't
readable by older versions of the code.  The new behavior is turned
off by default, however.  After turning on the new behavior, instead
of rolling back, you can turn it off again if you observe problems.
In order to turn on the new compaction code, the JVM system property
"tsd.feature.compactions" needs to be set.  You can set this by in
your `tsdb.local' file like so:
  JVMARGS='-Dtsd.feature.compactions'

With this change the TSD remembers every row it's writing to in HBase.
Periodically, it will read back old rows from HBase, and rewrite them
in a more efficient format.  The compacted form consists mostly in
concatenating together all the qualifiers and values inside that row.

When the TSD reads a row that isn't compacted properly (or at all)
it will fix it by overwriting the row again.  So the code is handling
properly weird cases such as data points coming in with very old
timestamps or out of order data etc.  It also handles properly races
in case multiple concurrent TSDs attempt to compact the same row at
the same time.

In order for this change to be truly effective, it's important to have
HBase compactions turned on, this way HBase can purge the individual
data points and keep the compacted form only.  If you wish to leave
HBase compactions turned off, consider triggering them periodically
(either via a cron job, or manually).

Note: `fsck' currently doesn't support the compacted rows, and will
emit a warning when it sees one.  Another commit will add support for
compacted rows in `fsck' in the future.

Change-Id: Ibc8f0154f7dd79d5337fa7effb645a297f9f94b2