Fix incorrect Cache-Control header for queries with relative dates.
When a query was using relative start/end date(s), the TSD was
incorrectly setting the `max-age' in the `Cache-Control' header
to 1 day if the end date ended up "far" in the past.  This is
incorrect as the `max-age' should depend on the time range covered
by the request for any query that uses relative dates.

This closes #10.

Change-Id: I885d45c120394409072b1719c1a09e8d56d77af6