Autotoolize OpenTSDB.
With this change, OpenTSDB switches to the common build process that
consists in running `./configure' then `make'.  For those building out
of a fresh Git copy, they would need to run `./bootstrap' once first.

Notable changes: The startup script, which used to be `./src/tsdb' is
now in the build directory, so it can be invoked simply with `./tsdb'.

Thanks to this change, OpenTSDB can be installed via `make install'
and de-installed via `make uninstall'.  This closes #32.

Change-Id: Ic40a57e67132578a6d0ee6375a1ca67d4e68b33e