Dynamically download dependencies.
Checking in the dependencies into Git was convenient, but the
dependencies are big and updating them is costly as every clone
of the repository will come with all the versions ever used for
each dependency.  Now we only keep the MD5 sum of the dependent
jars and we download them from Google Code.  Yes MD5 is cracked
etc, but the goal is simply to protect against corruption.

Dependencies get distributed in the tarball generated by `make dist'
so that means that only maintainers are expected to download them,
regular users who are going to build from a tarball won't have to
bother.  This has the important property that building from a release
will not require Internet connectivity.

Change-Id: Ic0e31af56caeb5ad083bd42a8e68720986b9114a