Fix `make check' by using the correct CLASSPATH to run tests.
Since c31d101 we find the script mygnuplot.sh through the CLASSPATH.
This means that unit tests need to be run with the src/ directory on
their CLASSPATH.  This revealed a latent problem, sometimes referred
to as the "Classloader hell".  Because src/ is now on the CLASSPATH,
src/logback.xml is found by logback, and the file is telling logback
to load some JMX stuff.  Various components fiddle with their custom
class loaders and conflict with PowerMock's, and pain ensues.  So we
now tell PowerMock to ignore various packages in order to avoid this
kind of errors:
  "LinkageError: loader constraint violation"
  "IncompatibleClassChangeError: Class X does not implement the
  requested interface Y"
  "ClassCastException:
  com.sun.org.apache.xerces.internal.jaxp.SAXParserFactoryImpl
  cannot be cast to javax.xml.parsers.SAXParserFactory"
And other similarly exciting exceptions.

Change-Id: I941cd5030f4e02fad239b3e0e5c12a3800ec4b47