Gracefully recover from expired scanners.
If a scanner times out during a read query, asynchbase would
transparently try to re-open the scanner, and this could lead
it to return a row that was already partially scanned.  Handle
this gracefully by detecting this condition, discarding what
we knew about this row prior to the timeout, and using the data
the new scanner returned as if it was the first time we scanned
that row.

Change-Id: I7f0173f184f3a30256712f85ac443f29c0db1d28