Only rebuild files that have changed.
Whenever a single `.java' file changed, the Makefile was rebuilding
everything.  This was because it's much faster to compile all the
`.java' files at once than one-by-one like in C/C++ projects.  But
when only a single file changes, it's still faster to recompile just
that file (or the set of files that changed).

With this change, editing a single file and compiling the project
now takes about twice less time and CPU cycles than previously.
Also the compilation time for single-file changes is now roughly
constant (depends on the size of the file) regardless of how many
files are in the project.

On my Mac this means I only wait 1.2 seconds vs >3 seconds.  Yay.

Change-Id: Id63bd483feb0ebd48134e6f704d1eda70c9e6921