Limit the number of concurrent flushes that can happen.
When the CompactionQueue contains, say, 1M rows, and flush() is called,
the code was attempting to compact all these rows concurrently.  This
is unrealistic and can either kill HBase and/or run the TSD out of RAM.
With this change, a single call to flush() can only initiate up to 10k
concurrent compactions, after which the code asynchronously waits until
the batch of 10k is done before flushing more.

Change-Id: I05ec3839e08f13455f80b3aba31acf5f42c7ac38