Don't flush the entire compaction queue on shutdown.
It's just too much work and most of it is unnecessary.  Here's why.
In a typical setup, you'd have at least 2 TSD to be N + 1.  If you
shutdown the first one, all the traffic will shift to the other one
by virtue of tcollectors reconnecting elsewhere.  This is going to
"fault in" almost all the rows that were in the compaction queue of
the TSD that got shutdown, so the other TSD will eventually compact
them.  It's possible that few rows will be left behind, and it's OK
since they can always be compacted later on subsequent reads.

Flushing the entire queue simply takes too long.  If there are, say,
600k rows sitting in the queue, and the TSD is able to flush 2500/s,
it would take 4 minutes to flush them all.  So instead we only flush
the ones we were really supposed to flush anyway, that is the rows
that are over 1h old.

Change-Id: I4594b714ebf0d831c9fb8d314a38154439166ec0