Licensing adjustment: allow LGPLv2.1+ instead of LGPLv3+.
Until now, OpenTSDB was available under the terms of the LGPL,
version 3 or any later version.  This change also allows the
version 2.1 of the LGPL license.  Those who already accepted
the terms of the LGPLv3 can keep doing so, because LGPLv2.1
also states that one can use any later version.

The net result of this change is that one can now use OpenTSDB
under the terms of LGPLv2.1 or GPLv2, if they so chose.

Change-Id: I67e0d47206db6a5544fdadc9c5fb9cd9466c1ad3