Redirect graph requests with no type to the home page.
When a graph request comes in with no `&png', `&json', or `&ascii', we
used to show a basic, static HTML page with an <img> of the graph in it.
This wasn't very useful.  Now that the home page can reload its state
from the URL fragment, redirect to it with the right fragment so that
the graph appears with the form pre-populated properly.  This way it
becomes very easy to edit a graph.  Take an URL to an image from TSD,
drop the `&png' from it, load it in your browser, and voilà, you can
edit someone else's graph super easily!

Change-Id: If8316aaf5126a2e9ca918ba044bdef65f39c6405