Fix regression that prevents reading pre-1.0 floating point values.
Before OpenTSDB 1.0, floating point values were incorrectly encoded
when stored in HBase.  Their flags pretended they were stored on 4
bytes, but they were instead stored on 8 bytes with 4 leading zeros.
This bug got fixed in f4f405c9, with a workaround added to detect
such values and fix them automatically as part of the compaction
process.

The code contains an optimization, however, to skip the compaction
for single-KeyValue rows.  This is because if a row contains only
one KeyValue, there is obviously no point in running it through the
compaction code, since there is nothing to compact.  But because
the compaction code was the one that was transparently fixing up
those incorrectly encoded floating point values, this optimization
prevented them from getting fixed up.

So add some code to detect this case and fix up the KeyValue anyway.

There was also another bug in the code dumping time series data, where
it was also ignoring this corner case and incorrectly parsing the value.

In both cases the end result for a user is that the code would pretend
the value is 0, when in fact it's not.

Thanks to Paula Keezer for reporting the bug.

Change-Id: I8cc8915602185a3085ad0228554494efdc1ba1e1