Don't block mouse click events on the image.
Doing so means that when the focus is on one of the text boxes,
clicking on the image doesn't remove the focus from the text box,
which is annoying as change of focus is often used to trigger an
update of the graph.

Also increase the minimum selection size to avoid unintentional
drag-to-zoom events, which happen easily with a touch pad.