Improve TSD write throughput by 10x.
This fixes an important performance bug in TSD.  The bug is triggered
when large TCP frames are handled by TSD's telnet-style decoding logic.
TSD is trying to split the input lines, however it was using a Netty
class that exhibits very poor performance for this job.  This due to
the fact Netty's DelimiterBasedFrameDecoder is built to handle multiple
delimiters (e.g. "\n" or "\r\n"), and it would always attempt to match
all delimiters to the input buffer to try to find the smallest match.

So on large input buffers, which is particularly worse when the client
is connected from localhost as the MTU of loopback interfaces tend to
be very large, the code would always attempt to look for a "\r\n" in
the entire input, thereby wasting a great deal of CPU cycles.

This code replaces Netty's DelimiterBasedFrameDecoder with a simpler
decoder that only supports splitting lines on "\n" or "\r\n" and
always works in one pass.  Single-socket throughput has been seen
to increase by as much as 20x, and overall throughput with several
clients contending for CPU time has been seen to increase around 10x.

This closes #120.