Add a workaround to prevent the graph from being dragged around.
On mouse-down events we need to prevent the event's default behavior as
it may make the browser start dragging the image around.  We don't want,
we want to handle the dragging ourselves for the drag-to-zoom box.

However preventing the event's default behavior has another obnoxious
consequence: it doesn't remove the focus from whatever had it previously,
which may cause the image to not get reloaded if a text field was being
changed and the user simply "clicked away" on the image.