Help forward compatibility by ignoring invalid KeyValues.
Right now all data points are written with 2-byte qualifiers.  When we
compact data points, we paste together the qualifiers, so we end up with
a multiple of 2 bytes no matter what.  So when compacting a row, only
consider those KeyValues that have a qualifier that is even and has 2 or
more bytes, and silently ignore the rest.