[check_tsd] Explicitly check for w/c is None
In its current form, if you want to warn between 0 and `x` and go
critical after `x` (i.e. `-w 0 -c 10`) the script will set `options.warning`
to `x` as `not 0` or `not 0.0` will evaluate to true. This changes it so that
`None` is explicitly checked for, thus if you pass `-w 0 -c 10` then
`options.warning` will remain `0` and `options.critical` will still be 10

Signed-off-by: Chris Larsen <clarsen@yahoo-inc.com>