Simplified deployment. With HBase accessible, and gnuplot installed, this works: java -jar ./target/opentsdb-2.1.0RC1-fat.jar tsd.
No requirement for a pre-deployed UI static content directory. The
FatJar contains all the static resources and unloads this content
into the configured/default directory at start time, including the
GWT Web App content and the gnuplot wrapper script. If any content
already exists and has a newer timestamp than the corresponding
resource in the FatJar, the resource is left as is. An additional
FatJar provided command line tool allows for the pre-creation of
the static content directory, which is useful in development.

The process PID is written to a default [and configurable] pid file,
and deleted on clean shutdown.

URL based config and include supporting the load of configuration
files from http:// or file:// based sources.

Configuration properties can be tokenized with System Prop and/or
Environmental Variable decodes as well as JavaScript snippets to
support dynamically computed property values according to the
deployment environment.

Refined logging configuration with configuration override-able
external logging config file location and command line options
to set logging levels for major packages in OpenTSDB and
associated libraries.

Signed-off-by: Chris Larsen <clarsen@yahoo-inc.com>