Start the migration to the new query pipeline wherin all of the components are QueryNode objects in the call graph. This greatly simplifies the code and should help to improve performance a bit. So far this implements the MockDataStore and a TSDBV2Pipeline that parses OpenTSDB 2.x queries (both regular and expressions, graphite style coming later).
Add the TimeSeriesDataSource interface for use by metric fetchers.
Rename QueryListener to QuerySink to be more descriptive.
Add the QueryContextBuilder interface, this is the Java API for creating
a query context for execution.
Add stubs in the registry for QueryNodeFactory(s).
Add SlicedTimeSeries to replace the sliced iterator. Still has some TODOs.
Rework the NumericMillisecondShard to fit the new pipeline.
Add AbstractQueryNode as a new base for nodes.
Add AbstractQueryPipelineContext as a new base for the pipeline context.
This does most of the query heavy lifting by storing and initializing the
call graph and routing results to the sinks given by the caller via the API.
It handles all modes of querying.
Add DefaultQueryContextBuilder as a default context builder. May have more
in the future.
Add QuerySourceConfig as a simple base time series data source config
implementation. More work for this in the future.
Remove the StorageQueryExecutor since we'll just use storage plugins directly.
Modify the TimeSeriesQuery POJO to store a map of the filters keyed by ID
so consumers don't have to iterate the whole bloody list each time they need
just a single filter.
TimeSeriesDataStore now requires an implementation of the QueryNodeFactory.
Disable a ton of tests till we fix up the executors to be nodes.